Update: Moved styling to bottom of components.Added new folders.
New folders were created to better organize components. Additionally, styling was moved to bottom of components to make JSX more readable.
