+++ src/actions/index.js
@@ -1,14 +1,2 @@
 export { LOGIN_START, LOGIN_SUCCESS, LOGIN_FAILURE, initiateLogin } from './login';
 export {PROD_STAT_START, PROD_STAT_SUCCESS, PROD_STAT_FAILURE, getProductStatData} from './dataProductStatsVue';
-// export {FETCH_DATA_START,
-//     FETCH_DATA_SUCCESS,
-//     FETCH_DATA_FAILURE,getData} from './getdata';
-
-// export {
-//   SEND_SENTIMENT_START,
-//   SEND_SENTIMENT_SUCCESS,
-//   SEND_SENTIMENT_FAILURE,
-//   sentimentSend
-// } from './sentimentSend';
-
-// export { REGISTER_START, REGISTER_SUCCESS, REGISTER_FAILURE, initiateRegister } from './register';
