+++ src/actions/login.js
@@ -21,7 +21,7 @@ export const initiateLogin = creds => dispatch =>{
     .post('https://tieme-ndo-backend.herokuapp.com/oauth/token',body, {
         headers: {
           "Content-Type": "application/x-www-form-urlencoded",
-          "Authorization": 'Basic ' + `${window.btoa(process.env.REACT_APP_AUTH_CLIENT_ID + ':' + process.env.REACT_APP_AUTH_CLIENT_SECRET)}`}
+          "Authorization": 'Basic ' + `${window.btoa('lambda-client' + ':' + 'lambda-secret')}`}
         //   process.env.REACT_APP_AUTH_CLIENT_ID + ':' + process.env.REACT_APP_AUTH_CLIENT_SECRET
         //   
         })
