+++ src/components/StatisticsDashboardComp/MultiGraphNav.js
@@ -4,6 +4,51 @@ import { NavLink } from 'react-router-dom'
 import styled, { css } from "styled-components";
 import { BrowserRouter as Router, Route, Link } from "react-router-dom";
 
+
+
+
+export default class MultiGraphNav extends Component{
+    constructor(props){
+        super(props);
+        this.state = {
+            question:true
+        }
+    }
+
+    componentDidMount(){
+     
+    }
+
+
+    questionClicked =()=>{
+        this.setState({question: !this.state.question})
+    }
+
+    render(){
+        return(
+            <StyledMultiGraphNav>
+                <div className ="info-div">
+                    <i  onClick={this.questionClicked} class="far fa-question-circle"></i>
+                    <h5 id={this.state.question ? 'hidden-stat-desc':''} className="extra-stat-detail">
+                    Statistics above based on totals  across all products or loans.</h5>
+                </div>
+                
+                <h4>Sales</h4>
+                <h3 className="multigraph-current-product">{"corn".toUpperCase()}</h3>
+                <div className="multigraph-nav">
+                    <NavLink className="navlink"><div className="multigraph-link"><h4>1D</h4></div></NavLink>
+                    <NavLink className="navlink"><div className="multigraph-link"><h4>1W</h4></div></NavLink>
+                    <NavLink className="navlink"><div className="multigraph-link"><h4>1M</h4></div></NavLink>
+                    <NavLink className="navlink"><div className="multigraph-link"><h4>1Y</h4></div></NavLink>
+
+                    
+                </div>
+
+            </StyledMultiGraphNav>
+        )
+    }
+}
+
 const sizes = {
     desktop: 992,
     tablet: 768,
@@ -111,64 +156,3 @@ const StyledMultiGraphNav = styled.div`
     }
 
 `
-
-
-export default class MultiGraphNav extends Component{
-    constructor(props){
-        super(props);
-        this.state = {
-            question:true
-        }
-    }
-
-    componentDidMount(){
-        // this.setHoverDetailListener();
-        // this.setMouseOutListener();
-    }
-
-    // setHoverDetailListener = ()=> {
-    //     var inputon = document.querySelectorAll(".menu-link");
-    //     inputon.forEach(item => item.addEventListener("mouseover", function(event) {
-    //     let mydesc = document.querySelector('#hidden-stat-desc');
-    //     mydesc.id ="visible";
-    //     }));
-    // }
-
-    // setMouseOutListener = ()=>{
-    //     var inputoff = document.querySelectorAll(".menu-link");
-    //     inputoff.forEach(item => item.removeEventListener("mouseout", function(event) {
-    //     let mydesc = document.querySelector('#visible');
-    //     mydesc.id ="visible";
-    //     }));
-
-    
-    // }
-
-    questionClicked =()=>{
-        this.setState({question: !this.state.question})
-    }
-
-    render(){
-        return(
-            <StyledMultiGraphNav>
-                <div className ="info-div">
-                    <i  onClick={this.questionClicked} class="far fa-question-circle"></i>
-                    <h5 id={this.state.question ? 'hidden-stat-desc':''} className="extra-stat-detail">
-                    Statistics above based on totals  across all products or loans.</h5>
-                </div>
-                
-                <h4>Sales</h4>
-                <h3 className="multigraph-current-product">{"corn".toUpperCase()}</h3>
-                <div className="multigraph-nav">
-                    <NavLink className="navlink"><div className="multigraph-link"><h4>1D</h4></div></NavLink>
-                    <NavLink className="navlink"><div className="multigraph-link"><h4>1W</h4></div></NavLink>
-                    <NavLink className="navlink"><div className="multigraph-link"><h4>1M</h4></div></NavLink>
-                    <NavLink className="navlink"><div className="multigraph-link"><h4>1Y</h4></div></NavLink>
-
-                    
-                </div>
-
-            </StyledMultiGraphNav>
-        )
-    }
-}
