+++ src/components/StatisticsDashboardComp/ProductStatsMiniChart.js
