+++ src/components/StatisticsDashboardComp/StatisticsVue.js
@@ -6,32 +6,7 @@ import ProductDescriptives from './ProductDescriptives';
 import ProductMultiVue from './ProductMultiVue';
 
 
-const sizes = {
-    desktop: 992,
-    tablet: 768,
-    phone: 576
-};
-  
-  const media = Object.keys(sizes).reduce((acc, label) => {
-    acc[label] = (...args) =>
-      css`
-        @media (max-width: ${sizes[label]}px) {
-          ${css(...args)}
-        }
-      `;
-  
-    return acc;
-}, {});
 
-const StyledStatisticsVue = styled.div`
-    height:100%;
-    width:100%;
-    display:flex;
-    justify-content:flex-end;
-    align-items:center;
-
-
-`
 
 
 
@@ -56,3 +31,31 @@ export default class StatisticsVue extends Component{
     }
 }
 
+//begin styling
+
+const sizes = {
+  desktop: 992,
+  tablet: 768,
+  phone: 576
+};
+
+const media = Object.keys(sizes).reduce((acc, label) => {
+  acc[label] = (...args) =>
+    css`
+      @media (max-width: ${sizes[label]}px) {
+        ${css(...args)}
+      }
+    `;
+
+  return acc;
+}, {});
+
+const StyledStatisticsVue = styled.div`
+  height:100%;
+  width:100%;
+  display:flex;
+  justify-content:flex-end;
+  align-items:center;
+
+
+`
